added missing data and adjusted some types to be S3 accurate

+++ Arrowgene.Ddon.Shared/Entity/Structure/CDataEquipSlot.cs
@@ -10,8 +10,8 @@ namespace Arrowgene.Ddon.Shared.Entity.Structure
         }
         public int Unk0 { get; set; }
         public int Unk1 { get; set; }
-        public int Unk2 { get; set; }
-        public int Unk3 { get; set; }
+        public byte Unk2 { get; set; }
+        public ushort Unk3 { get; set; }
 
         public class Serializer : EntitySerializer<CDataEquipSlot>
         {
@@ -19,8 +19,8 @@ namespace Arrowgene.Ddon.Shared.Entity.Structure
             {
                 WriteInt32(buffer, obj.Unk0);
                 WriteInt32(buffer, obj.Unk1);
-                WriteInt32(buffer, obj.Unk2);
-                WriteInt32(buffer, obj.Unk3);
+                WriteByte(buffer, obj.Unk2);
+                WriteUInt16(buffer, obj.Unk3);
             }
 
             public override CDataEquipSlot Read(IBuffer buffer)
@@ -28,8 +28,8 @@ namespace Arrowgene.Ddon.Shared.Entity.Structure
                     CDataEquipSlot obj = new CDataEquipSlot();
                     obj.Unk0 = ReadInt32(buffer);
                     obj.Unk1 = ReadInt32(buffer);
-                    obj.Unk2 = ReadInt32(buffer);
-                    obj.Unk3 = ReadInt32(buffer);
+                    obj.Unk2 = ReadByte(buffer);
+                    obj.Unk3 = ReadUInt16(buffer);
                     return obj;
                 
             }
